Added check button for checking off completed buckets and protocol for communication between cell and view controller

+++ JustMovdApp/FoursquareVenue.h
@@ -19,7 +19,7 @@
 @property (strong, nonatomic) NSString     *postalCode;
 @property (strong, nonatomic) NSString     *city;
 @property (strong, nonatomic) UIImage      *fullImage;
-@property (nonatomic) float lat;
-@property (nonatomic) float lng;
+@property (nonatomic) double lat;
+@property (nonatomic) double lng;
 
 @end
